requirements.txt
numpy>=1.24.0
Pillow>=10.0.0
pdf2image>=1.16.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
